--- src/main/java/com/example/server/HomeController.java
@@ -1,20 +1,0 @@
-package com.example.server;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
-
-@RestController
-public class HomeController {
-	@GetMapping({"", "/"})
-	public String home() {
-		return "Dasibom Home Controller";
-	}
-
-	@GetMapping("/social/login")
-	public ModelAndView login() {
-		ModelAndView mv = new ModelAndView();
-		mv.setViewName("home");
-		return mv;
-	}
-}
