+++ src/main/java/com/example/server/domain/member/application/MemberFindService.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.server.domain.member.model.Member;
+import com.example.server.domain.member.persistence.MemberQueryRepository;
 import com.example.server.global.exception.BusinessException;
 import com.example.server.global.exception.ErrorCode;
-import com.example.server.domain.member.persistence.MemberQueryRepository;
 
 import lombok.RequiredArgsConstructor;
