+++ src/main/java/com/example/server/domain/member/model/Member.java
@@ -17,9 +17,9 @@ import javax.persistence.Table;
 
 import org.hibernate.annotations.CreationTimestamp;
 
-import com.example.server.global.config.oauth.provider.OAuth2Provider;
 import com.example.server.domain.image.model.Image;
 import com.example.server.domain.member.model.constants.RoleType;
+import com.example.server.global.oauth.provider.OAuth2Provider;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
