+++ src/main/java/com/example/server/domain/member/persistence/MemberQueryRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.server.global.config.oauth.provider.OAuth2Provider;
 import com.example.server.domain.member.model.Member;
+import com.example.server.global.oauth.provider.OAuth2Provider;
 
 @Repository
 @Transactional(readOnly = true)
