--- src/main/java/com/example/server/global/config/auth/PrincipalDetails.java
@@ -1,79 +1,0 @@
-package com.example.server.global.config.auth;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import com.example.server.domain.member.model.Member;
-
-public class PrincipalDetails implements UserDetails, OAuth2User {
-
-	private static final long serialVersionUID = 1L;
-	private Member user;
-	private Map<String, Object> attributes;
-
-	public PrincipalDetails(Member user, Map<String, Object> attributes) {
-		this.user = user;
-		this.attributes = attributes;
-	}
-
-	public Member getUser() {
-		return user;
-	}
-
-	@Override
-	public String getPassword() {
-		return user.getPassword();
-	}
-
-	@Override
-	public String getUsername() {
-		return user.getUsername();
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		Collection<GrantedAuthority> collet = new ArrayList<GrantedAuthority>();
-		collet.add(() -> {
-			return user.getRole().toString();
-		});
-		return collet;
-	}
-
-	// 리소스 서버로 부터 받는 회원정보
-	@Override
-	public Map<String, Object> getAttributes() {
-		return attributes;
-	}
-
-	// User, Admin 의 PrimaryKey
-	@Override
-	public String getName() {
-		return user.getId() + "";
-	}
-
-}
