--- src/main/java/com/example/server/global/config/jwt/JwtHeaderUtil.java
@@ -1,23 +1,0 @@
-package com.example.server.global.config.jwt;
-
-import javax.servlet.http.HttpServletRequest;
-
-public class JwtHeaderUtil {
-
-	private final static String HEADER_AUTHORIZATION = "Authorization";
-	private final static String TOKEN_PREFIX = "Bearer ";
-
-	public static String getAccessToken(HttpServletRequest request) {
-		String headerValue = request.getHeader(HEADER_AUTHORIZATION);
-
-		if (headerValue == null) {
-			return null;
-		}
-
-		if (headerValue.startsWith(TOKEN_PREFIX)) {
-			return headerValue.substring(TOKEN_PREFIX.length());
-		}
-
-		return null;
-	}
-}
