--- src/main/java/com/example/server/global/config/jwt/dto/AccessToken.java
@@ -1,47 +1,0 @@
-package com.example.server.global.config.jwt.dto;
-
-import java.security.Key;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@RequiredArgsConstructor
-public class AccessToken {
-
-	@Getter
-	private final String token;
-
-	private final Key key;
-
-	public boolean validate() {
-		return this.getTokenClaims() != null;
-	}
-
-	public Claims getTokenClaims() {
-		try {
-			return Jwts.parserBuilder()
-				.setSigningKey(key)
-				.build()
-				.parseClaimsJws(token)
-				.getBody();
-		} catch (SecurityException e) {
-			log.info("Invalid JWT signature.");
-		} catch (MalformedJwtException e) {
-			log.info("Invalid JWT token.");
-		} catch (ExpiredJwtException e) {
-			log.info("Expired JWT token.");
-		} catch (UnsupportedJwtException e) {
-			log.info("Unsupported JWT token.");
-		} catch (IllegalArgumentException e) {
-			log.info("JWT token compact of handler are invalid.");
-		}
-		return null;
-	}
-}
