--- src/main/java/com/example/server/global/config/jwt/dto/AuthResponse.java
@@ -1,16 +1,0 @@
-package com.example.server.global.config.jwt.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class AuthResponse {
-	private String accessToken;
-	private String refreshToken;
-	private Boolean isNewMember;
-}
