--- src/main/java/com/example/server/global/config/jwt/dto/RefreshToken.java
@@ -1,15 +1,0 @@
-package com.example.server.global.config.jwt.dto;
-
-import javax.persistence.Id;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class RefreshToken {
-
-	@Id
-	private String refreshToken;
-	private String memberId; // socialId
-}
