--- src/main/java/com/example/server/global/config/jwt/exception/TokenValidFailedException.java
@@ -1,12 +1,0 @@
-package com.example.server.global.config.jwt.exception;
-
-public class TokenValidFailedException extends RuntimeException {
-
-	public TokenValidFailedException() {
-		super("Failed to generate Token.");
-	}
-
-	public TokenValidFailedException(String message) {
-		super(message);
-	}
-}
