--- src/main/java/com/example/server/global/config/oauth/provider/NaverUserInfo.java
@@ -1,45 +1,0 @@
-package com.example.server.global.config.oauth.provider;
-
-import java.util.Map;
-import java.util.Optional;
-
-public class NaverUserInfo extends OAuth2UserInfo {
-
-	public NaverUserInfo(Map<String, Object> attributes) {
-		super(attributes);
-	}
-
-	@Override
-	public String getProviderId() {
-		return Optional.ofNullable(attributes.get("response"))
-			.map(account -> (String)((Map<String, Object>)account).get("id"))
-			.orElse(null);
-	}
-
-	@Override
-	public OAuth2Provider getProvider() {
-		return OAuth2Provider.NAVER;
-	}
-
-	@Override
-	public String getEmail() {
-		return Optional.ofNullable(attributes.get("response"))
-			.map(account -> (String)((Map<String, Object>)account).get("email"))
-			.orElse(null);
-	}
-
-	@Override
-	public String getName() {
-		return Optional.ofNullable(attributes.get("response"))
-			.map(account -> (String)((Map<String, Object>)account).get("name"))
-			.orElse(null);
-	}
-
-	@Override
-	public String getMobile() {
-		return Optional.ofNullable(attributes.get("response"))
-			.map(account -> (String)((Map<String, Object>)account).get("mobile"))
-			.orElse(null);
-	}
-
-}
