+++ src/main/java/com/example/server/global/jwt/dto/RefreshToken.java
@@ -1,0 +1,15 @@
+package com.example.server.global.jwt.dto;
+
+import javax.persistence.Id;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class RefreshToken {
+
+	@Id
+	private String refreshToken;
+	private String memberId; // socialId
+}
