+++ src/main/java/com/example/server/global/jwt/exception/TokenValidFailedException.java
@@ -1,0 +1,12 @@
+package com.example.server.global.jwt.exception;
+
+public class TokenValidFailedException extends RuntimeException {
+
+	public TokenValidFailedException() {
+		super("Failed to generate Token.");
+	}
+
+	public TokenValidFailedException(String message) {
+		super(message);
+	}
+}
