+++ src/main/java/com/example/server/global/jwt/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,0 +1,30 @@
+package com.example.server.global.jwt.handler;
+
+import static com.example.server.global.exception.ErrorCode.*;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+
+import com.example.server.global.exception.BusinessException;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Component
+public class OAuth2AuthenticationFailureHandler implements AuthenticationFailureHandler {
+	@Override
+	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException exception) throws IOException, ServletException {
+		//        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+		//        response.getWriter().write("소셜 로그인 실패! 서버 로그를 확인해주세요.");
+		log.info("소셜 로그인에 실패했습니다. 에러 메시지 : {}", exception.getMessage());
+		throw new BusinessException(SOCIAL_LOGIN_ERROR);
+	}
+}
